Move eventPairFromPersistenceIncremental to Hydra.Events.FileBased
This implementation and its tests will serve as an example for an
EventSource and EventSink.
