+++ hydra-node/hydra-node.cabal
@@ -69,6 +69,7 @@ library
     Hydra.Crypto
     Hydra.Environment
     Hydra.Events
+    Hydra.Events.FileBased
     Hydra.HeadId
     Hydra.HeadLogic
     Hydra.HeadLogic.Error
@@ -292,6 +293,7 @@ test-suite tests
     Hydra.Chain.Direct.WalletSpec
     Hydra.ContestationPeriodSpec
     Hydra.CryptoSpec
+    Hydra.Events.FileBasedSpec
     Hydra.FireForgetSpec
     Hydra.HeadLogicSnapshotSpec
     Hydra.HeadLogicSpec
