+++ hydra-node/test/Hydra/PersistenceSpec.hs
@@ -8,10 +8,8 @@ import Test.Hydra.Prelude
 import Data.Aeson (Value (..))
 import Data.Aeson qualified as Aeson
 import Data.Text qualified as Text
-import Hydra.Events (EventSink (..), EventSource (..), StateEvent (..), getEvents, putEvent)
-import Hydra.Ledger.Simple (SimpleTx)
-import Hydra.Persistence (Persistence (..), PersistenceException (..), PersistenceIncremental (..), createPersistence, createPersistenceIncremental, eventPairFromPersistenceIncremental)
-import Test.QuickCheck (checkCoverage, cover, elements, forAllShrink, ioProperty, oneof, sublistOf, suchThat, (===))
+import Hydra.Persistence (Persistence (..), PersistenceException (..), PersistenceIncremental (..), createPersistence, createPersistenceIncremental)
+import Test.QuickCheck (checkCoverage, cover, elements, oneof, suchThat, (===))
 import Test.QuickCheck.Gen (listOf)
 import Test.QuickCheck.Monadic (monadicIO, monitor, pick, run)
 
@@ -73,68 +71,6 @@ spec = do
               `shouldThrow` \case
                 IncorrectAccessException{} -> True
                 _ -> False
-
-  describe "eventPairFromPersistenceIncremental" $ do
-    prop "can handle continuous events" $
-      forAllShrink genContinuousEvents shrink $ \events ->
-        ioProperty $ do
-          withEventSourceAndSink $ \EventSource{getEvents} EventSink{putEvent} -> do
-            forM_ events putEvent
-            loadedEvents <- getEvents
-            pure $
-              loadedEvents === events
-
-    prop "can handle non-continuous events" $
-      forAllShrink (sublistOf =<< genContinuousEvents) shrink $ \events ->
-        ioProperty $ do
-          withEventSourceAndSink $ \EventSource{getEvents} EventSink{putEvent} -> do
-            forM_ events putEvent
-            loadedEvents <- getEvents
-            pure $
-              loadedEvents === events
-
-    prop "can handle duplicate events" $
-      forAllShrink genContinuousEvents shrink $ \events ->
-        ioProperty $
-          withEventSourceAndSink $ \EventSource{getEvents} EventSink{putEvent} -> do
-            -- Put some events
-            forM_ events putEvent
-            loadedEvents <- getEvents
-            -- Put the loaded events again (as the node would do)
-            forM_ loadedEvents putEvent
-            allEvents <- getEvents
-            pure $
-              allEvents === loadedEvents
-
-    prop "can bootstrap from plain StateChanged events" $
-      forAllShrink genContinuousEvents shrink $ \events -> do
-        ioProperty $ do
-          withTempDir "hydra-persistence" $ \tmpDir -> do
-            -- Store state changes directly (legacy)
-            let stateChanges = map stateChanged events
-            PersistenceIncremental{append} <- createPersistenceIncremental (tmpDir <> "/data")
-            forM_ stateChanges append
-            -- Load and store events through the event source interface
-            (EventSource{getEvents}, EventSink{putEvent}) <-
-              eventPairFromPersistenceIncremental
-                =<< createPersistenceIncremental (tmpDir <> "/data")
-            loadedEvents <- getEvents
-            -- Store all loaded events like the node would do
-            forM_ loadedEvents putEvent
-            pure $
-              map stateChanged loadedEvents === stateChanges
-
-genContinuousEvents :: Gen [StateEvent SimpleTx]
-genContinuousEvents =
-  zipWith StateEvent [0 ..] <$> listOf arbitrary
-
-withEventSourceAndSink :: (EventSource (StateEvent SimpleTx) IO -> EventSink (StateEvent SimpleTx) IO -> IO b) -> IO b
-withEventSourceAndSink action =
-  withTempDir "hydra-persistence" $ \tmpDir -> do
-    (eventSource, eventSink) <-
-      eventPairFromPersistenceIncremental
-        =<< createPersistenceIncremental (tmpDir <> "/data")
-    action eventSource eventSink
 
 genPersistenceItem :: Gen Aeson.Value
 genPersistenceItem =
